refactor(specs): update feature branch references to dev-junior
Updated all feature specification files and templates to reference the new branch name "dev-junior" instead of "001-login-google". This includes changes to README, data-model, feature-spec, quickstart, and spec files in the specs/001-login-google directory, as well as corresponding template files. The migration reference was also removed from the quickstart documentation to align with the updated structure.